Added Hardware Acceleration option
Added hardware acceleration option to settings menu
Updated config file read, save, update

updated folder checking to be more  robust due to error found

+++ FolderCleanup.cs
@@ -35,13 +35,19 @@ namespace xPlatformLukma
         public string CleanUpFolder() 
         {
             string sErrors = "";
-            List<string> fileList = SearchFoldersForFiles(_baseFolder, _dayCount, _currentDate);
-
-            if (fileList.Count > 0)
+            if (Directory.Exists(_baseFolder))
             {
-                sErrors = DeleteOldFiles(fileList);
-            }
+                List<string> fileList = SearchFoldersForFiles(_baseFolder, _dayCount, _currentDate);
 
+                if (fileList.Count > 0)
+                {
+                    sErrors = DeleteOldFiles(fileList);
+                }
+            }
+            else 
+            {
+                sErrors = "Files cannot be cleaned up. Folder doesn't exist: " + _baseFolder;
+            }
 
             return sErrors;
         }
